newHeader renamed to Header and now serves as default
