+++ src/components/Header/Navbar/Menu/DropdownList/Platform.js
