+++ src/components/index.js
@@ -1,3 +1,4 @@
+export { default as Header } from './Header'
 export { default as Footer } from './Footer'
 export { default as AppContainer } from './AppContainer'
 export { default as NavLink } from './NavLink'
@@ -9,7 +10,6 @@ export { default as WhiteButton } from './WhiteButton'
 export { default as LinkWithoutStyle } from './LinkWithoutStyle'
 export { default as NewsLetterForm } from './NewsLetterForm'
 export { default as AppContainerNew } from './AppContainerNew'
-export { default as HeaderNew } from './HeaderNew'
 export { default as FooterNew } from './FooterNew'
 export { default as WhiteButtonNew } from './WhiteButtonNew'
 export { default as NewsLetterFormNew } from './NewNewsLetterForm'
